Temperature Conversion
Added the .html and .js files. Coded the .html file to include a basic
title and description of what is happening and what the code is
running. Inside the .js file, I’ve coded a calculator that turns either
Fahrenheit to Celsius or the other way around. It then prints to the
console.